requirements.txt
alembic==1.12.1
azure-common==1.1.28
azure-nspkg==3.0.2
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
Click==8.1.7
cryptography==42.0.3
Flask==2.2.5
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.2.4
MarkupSafe==2.1.5
msal==1.26.0
pip==24.0
pycparser==2.21
pyodbc==4.0.39
pyopenssl==24.0.0
python-dateutil==2.8.2
python-editor==1.0.4
requests==2.31.0
setuptools==68.0.0
six==1.16.0
SQLAlchemy==2.0.27
urllib3==2.0.7
Werkzeug==2.2.3
WTForms==3.0.1
python-dotenv==0.21.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure_keyvault_secrets==4.7.0
